# restful/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/geo/address": {
            "post": {
                "description": "Retrieves geolocation data for a given address using the DaData API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geolocation"
                ],
                "summary": "Get geolocation data for an address",
                "parameters": [
                    {
                        "description": "Address",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Geolocation data retrieved",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.Address"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Address": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "Район в регионе",
                    "type": "string"
                },
                "area_fias_id": {
                    "description": "Код ФИАС района в регионе",
                    "type": "string"
                },
                "area_kladr_id": {
                    "description": "Код КЛАДР района в регионе",
                    "type": "string"
                },
                "area_type": {
                    "description": "Тип района в регионе (сокращенный)",
                    "type": "string"
                },
                "area_type_full": {
                    "description": "Тип района в регионе",
                    "type": "string"
                },
                "area_with_type": {
                    "description": "Район в регионе с типом",
                    "type": "string"
                },
                "beltway_distance": {
                    "description": "Расстояние от кольцевой в км.",
                    "type": "string"
                },
                "beltway_hit": {
                    "description": "Внутри кольцевой?",
                    "type": "string"
                },
                "block": {
                    "description": "Корпус/строение",
                    "type": "string"
                },
                "block_type": {
                    "description": "Тип корпуса/строения (сокращенный)",
                    "type": "string"
                },
                "block_type_full": {
                    "description": "Тип корпуса/строения",
                    "type": "string"
                },
                "capital_marker": {
                    "description": "Статус центра",
                    "type": "string"
                },
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "city_area": {
                    "description": "Административный округ (только для Москвы)",
                    "type": "string"
                },
                "city_district": {
                    "description": "Район города",
                    "type": "string"
                },
                "city_district_fias_id": {
                    "description": "Код ФИАС района города (заполняется, только если район есть в ФИАС)",
                    "type": "string"
                },
                "city_district_kladr_id": {
                    "description": "Код КЛАДР района города (не заполняется)",
                    "type": "string"
                },
                "city_district_type": {
                    "description": "Тип района города (сокращенный)",
                    "type": "string"
                },
                "city_district_type_full": {
                    "description": "Тип района города",
                    "type": "string"
                },
                "city_district_with_type": {
                    "description": "Район города с типом",
                    "type": "string"
                },
                "city_fias_id": {
                    "description": "Код ФИАС города",
                    "type": "string"
                },
                "city_kladr_id": {
                    "description": "Код КЛАДР города",
                    "type": "string"
                },
                "city_type": {
                    "description": "Тип города (сокращенный)",
                    "type": "string"
                },
                "city_type_full": {
                    "description": "Тип города",
                    "type": "string"
                },
                "city_with_type": {
                    "description": "Город с типом",
                    "type": "string"
                },
                "country": {
                    "description": "Страна",
                    "type": "string"
                },
                "country_iso_code": {
                    "description": "ISO-код страны",
                    "type": "string"
                },
                "entrance": {
                    "description": "Подъезд",
                    "type": "string"
                },
                "federal_district": {
                    "description": "Федеральный округ",
                    "type": "string"
                },
                "fias_actuality_state": {
                    "description": "Признак актуальности адреса в ФИАС",
                    "type": "string"
                },
                "fias_code": {
                    "description": "Иерархический код адреса в ФИАС (СС+РРР+ГГГ+ППП+СССС+УУУУ+ДДДД)",
                    "type": "string"
                },
                "fias_id": {
                    "description": "Код ФИАС",
                    "type": "string"
                },
                "fias_level": {
                    "description": "Уровень детализации, до которого адрес найден в ФИАС",
                    "type": "string"
                },
                "flat": {
                    "description": "Квартира",
                    "type": "string"
                },
                "flat_area": {
                    "description": "Площадь квартиры",
                    "type": "string"
                },
                "flat_cadnum": {
                    "description": "Кадастровый номер квартиры (22.4+). Заполняется в зависимости от тарифа «Дадаты».",
                    "type": "string"
                },
                "flat_fias_id": {
                    "description": "ФИАС-код квартиры",
                    "type": "string"
                },
                "flat_price": {
                    "description": "Рыночная стоимость квартиры",
                    "type": "string"
                },
                "flat_type": {
                    "description": "Тип квартиры (сокращенный)",
                    "type": "string"
                },
                "flat_type_full": {
                    "description": "Тип квартиры",
                    "type": "string"
                },
                "floor": {
                    "description": "Этаж",
                    "type": "string"
                },
                "geo_lat": {
                    "description": "Координаты: широта",
                    "type": "string"
                },
                "geo_lon": {
                    "description": "Координаты: долгота",
                    "type": "string"
                },
                "house": {
                    "description": "Дом",
                    "type": "string"
                },
                "house_cadnum": {
                    "description": "Кадастровый номер дома (22.4+). Заполняется в зависимости от тарифа «Дадаты».",
                    "type": "string"
                },
                "house_fias_id": {
                    "description": "Код ФИАС дома",
                    "type": "string"
                },
                "house_kladr_id": {
                    "description": "Код КЛАДР дома",
                    "type": "string"
                },
                "house_type": {
                    "description": "Тип дома (сокращенный)",
                    "type": "string"
                },
                "house_type_full": {
                    "description": "Тип дома",
                    "type": "string"
                },
                "kladr_id": {
                    "description": "Код КЛАДР",
                    "type": "string"
                },
                "metro": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.Metro"
                    }
                },
                "okato": {
                    "description": "Код ОКАТО",
                    "type": "string"
                },
                "oktmo": {
                    "description": "Код ОКТМО",
                    "type": "string"
                },
                "postal_box": {
                    "description": "Абонентский ящик",
                    "type": "string"
                },
                "postal_code": {
                    "description": "Индекс",
                    "type": "string"
                },
                "qc": {
                    "description": "Код качества"
                },
                "qc_complete": {
                    "description": "Код полноты"
                },
                "qc_geo": {
                    "description": "QualityCodeGeoRaw для clean вызовов он int для suggest в адресе банков он string поэтому в поле поставил interface{} чтобы работало и там и там)\\"
                },
                "qc_house": {
                    "description": "Код проверки дома"
                },
                "region": {
                    "description": "Регион",
                    "type": "string"
                },
                "region_fias_id": {
                    "description": "Код ФИАС региона",
                    "type": "string"
                },
                "region_iso_code": {
                    "description": "ISO-код региона",
                    "type": "string"
                },
                "region_kladr_id": {
                    "description": "Код КЛАДР региона",
                    "type": "string"
                },
                "region_type": {
                    "description": "Тип региона (сокращенный)",
                    "type": "string"
                },
                "region_type_full": {
                    "description": "Тип региона",
                    "type": "string"
                },
                "region_with_type": {
                    "description": "Регион с типом",
                    "type": "string"
                },
                "result": {
                    "description": "Стандартизованный адрес одной строкой",
                    "type": "string"
                },
                "settlement": {
                    "description": "Населенный пункт",
                    "type": "string"
                },
                "settlement_fias_id": {
                    "description": "Код ФИАС нас. пункта",
                    "type": "string"
                },
                "settlement_kladr_id": {
                    "description": "Код КЛАДР нас. пункта",
                    "type": "string"
                },
                "settlement_type": {
                    "description": "Тип населенного пункта (сокращенный)",
                    "type": "string"
                },
                "settlement_type_full": {
                    "description": "Тип населенного пункта",
                    "type": "string"
                },
                "settlement_with_type": {
                    "description": "Населенный пункт с типом",
                    "type": "string"
                },
                "source": {
                    "description": "Исходный адрес одной строкой",
                    "type": "string"
                },
                "square_meter_price": {
                    "description": "Рыночная стоимость м²",
                    "type": "string"
                },
                "street": {
                    "description": "Улица",
                    "type": "string"
                },
                "street_fias_id": {
                    "description": "Код ФИАС улицы",
                    "type": "string"
                },
                "street_kladr_id": {
                    "description": "Код КЛАДР улицы",
                    "type": "string"
                },
                "street_type": {
                    "description": "Тип улицы (сокращенный)",
                    "type": "string"
                },
                "street_type_full": {
                    "description": "Тип улицы",
                    "type": "string"
                },
                "street_with_type": {
                    "description": "Улица с типом",
                    "type": "string"
                },
                "tax_office": {
                    "description": "Код ИФНС для физических лиц",
                    "type": "string"
                },
                "tax_office_legal": {
                    "description": "Код ИФНС для организаций",
                    "type": "string"
                },
                "timezone": {
                    "description": "Часовой пояс",
                    "type": "string"
                },
                "unparsed_parts": {
                    "description": "Нераспознанная часть адреса. Для адреса",
                    "type": "string"
                }
            }
        },
...
        "handler.Metro": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "line": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}